# back/docs/swagger.yaml
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  models.Exercici:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      exercicisRutina:
        description: ExercicisRutina es una lista de rutinas que utilizan este exercici.
        items:
          $ref: '#/definitions/models.ExerciciRutina'
        type: array
      id:
        type: integer
      nom:
        description: |-
          Nom es el nombre del exercici.
          required: true
          example: Extensión de tríceps en polea
        type: string
      updatedAt:
        type: string
    type: object
...
  models.ExerciciRutina:
    properties:
      cicle:
        type: integer
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      diaRutina:
        type: integer
      exercici:
        $ref: '#/definitions/models.Exercici'
      exerciciID:
        type: integer
      id:
        type: integer
      numRepes:
        type: integer
      numSeries:
        type: integer
      ordre:
        type: integer
      percentatgeRM:
        type: integer
      rutinaID:
        type: integer
      updatedAt:
        type: string
    type: object
  models.ExerciciRutinaInput:
    properties:
      cicle:
        type: integer
      diaRutina:
        type: integer
      exerciciID:
        type: integer
      numRepes:
        type: integer
      numSeries:
        type: integer
      ordre:
        type: integer
      percentatgeRM:
        type: integer
    type: object
  models.ExerciciRutinaResponse:
    properties:
      cicle:
        type: integer
      diaRutina:
        type: integer
      exerciciID:
        type: integer
      id:
        type: integer
      nom:
        type: string
      numRepes:
        type: integer
      numSeries:
        type: integer
      ordre:
        type: integer
      percentatgeRM:
        type: integer
    type: object
...
  models.Rutina:
    properties:
      cicles:
        type: integer
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      descripcio:
        type: string
      diesDuracio:
        type: integer
      entrenadorID:
        type: integer
      exercicisRutina:
        items:
          $ref: '#/definitions/models.ExerciciRutina'
        type: array
      id:
        type: integer
      nom:
        type: string
      publica:
        type: boolean
      updatedAt:
        type: string
    type: object
  models.RutinaInput:
    properties:
      cicles:
        type: integer
      descripcio:
        type: string
      diesDuracio:
        type: integer
      exercicis:
        items:
          $ref: '#/definitions/models.ExerciciRutinaInput'
        type: array
      nom:
        type: string
    type: object
  models.RutinaResponse:
    properties:
      cicles:
        type: integer
      diesDuracio:
        type: integer
      exercicis:
        items:
          $ref: '#/definitions/models.ExerciciRutinaResponse'
        type: array
      id:
        type: integer
      nom:
        type: string
    type: object
...
  models.SuccessResponse:
    properties:
      data: {}
    type: object
...
  models.UsuariRutinaInput:
    properties:
      alumneID:
        type: integer
      rutinaID:
        type: integer
    type: object
info:
  contact: {}
paths:
...
  /api/rutines:
    get:
      consumes:
      - application/json
      description: Retrieves all the rutines from the database.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.RutinaResponse'
                  type: array
              type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get all rutines
      tags:
      - Rutines
    post:
      consumes:
      - application/json
      description: Creates a new rutine in the database.
      parameters:
      - description: Rutine to create
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.RutinaInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.RutinaResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Create a new rutine
      tags:
      - Rutines
  /api/rutines/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a rutine from the database.
      parameters:
      - description: ID of the rutine to delete
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.SuccessResponse'
            - properties:
                data:
                  type: string
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Delete a rutine
      tags:
      - Rutines
    put:
      consumes:
      - application/json
      description: Updates a rutine in the database.
      parameters:
      - description: ID of the rutine to update
        in: path
        name: id
        required: true
        type: integer
      - description: Rutine to update
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.RutinaInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.RutinaResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Update a rutine
      tags:
      - Rutines
  /api/rutines/{id}/canviarVisibilitat:
    put:
      consumes:
      - application/json
      description: Changes the visibility of a rutine in the database.
      parameters:
      - description: ID of the rutine to change the visibility
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Rutina'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Change the visibility of a rutine
      tags:
      - Rutines
  /api/rutines/assignarRutina:
    post:
      consumes:
      - application/json
      description: Assigns a rutine to an alumne in the database.
      parameters:
      - description: Alumne to assign the rutine to
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.UsuariRutinaInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.SuccessResponse'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Assign a rutine to an alumne
      tags:
      - Rutines
  /api/rutines/rutinesEntrenador:
    get:
      consumes:
      - application/json
      description: Retrieves all the rutines of an entrenador from the database.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.RutinaResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Get all rutines of an entrenador
      tags:
      - Rutines
  /api/rutines/rutinesPubliques:
    get:
      consumes:
      - application/json
      description: Retrieves all the public rutines from the database.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.RutinaResponse'
                  type: array
              type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get all public rutines
      tags:
      - Rutines
...
securityDefinitions:
  Bearer:
    description: 'Enter the token with the `Bearer: ` prefix, e.g. "Bearer abcde12345".'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"